+++ ichnaea/api/locate/country/views.py
@@ -1,4 +1,3 @@
-from ichnaea.api.locate.searcher import CountrySearcher
 from ichnaea.api.locate.locate_v2.views import LocateV2View
 
 
@@ -7,7 +6,7 @@ class CountryView(LocateV2View):
     # TODO: Disable API key checks and logging, for the initial wave
     check_api_key = False
     error_on_invalidkey = False
-    searcher = CountrySearcher
+    searcher = 'country_searcher'
     route = '/v1/country'
     view_name = 'country'